pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "duperscooper"
version = "0.1.0"
description = "Duplicate audio file finder"
readme = "README.md"
requires-python = ">=3.8"
license = {file = "LICENSE"}
dependencies = [
    "pydub>=0.25.1",
    "imagehash>=4.3.1",
    "pillow>=10.0.0",
    "tqdm>=4.66.0",
]

[project.scripts]
duperscooper = "duperscooper.__main__:main"

[tool.black]
line-length = 88
target-version = ['py38', 'py39', 'py310', 'py311']

[tool.ruff]
line-length = 88
target-version = "py38"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = []
fixable = ["ALL"]
unfixable = []

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
    "slow: marks tests as slow running",
]